save_image.py
from PIL import Image
import io

# Create a new image with the table data
def create_table_image():
    # This would typically come from the PDF, but for now we'll create a placeholder
    # that indicates where the table data should be loaded from
    img = Image.new('RGB', (800, 600), color='white')
    img.save('financial_data.jpg') 