requirements.txt
torch==2.2.0
transformers==4.37.2
sentence-transformers==2.5.1
streamlit==1.30.0
pandas==2.2.0
python-dotenv==1.0.0
tabula-py==2.9.0
PyPDF2==3.0.1
openpyxl==3.1.2
pillow==10.0.0
pdf2image==1.17.0
opencv-python==4.9.0.80 